improving time of federated round

+++ workflows/airflow-components/plugins/kaapana/operators/HelperCaching.py
@@ -88,6 +88,7 @@ def cache_operator_output(func):
             raise AssertionError("Invalid name '{}' for manage_cache. It must be set to None, 'ignore', 'cache', 'overwrite' or 'clear'".format(self.manage_cache))
 
         run_id, dag_run_dir, dag_run, downstream_tasks = get_operator_properties(*args, **kwargs)
+        downstream_tasks_ids = [task.task_id for task in downstream_tasks]
         conf = dag_run.conf
         if conf is not None and 'client_job_id' in conf:
             update_job(conf['client_job_id'], status='running', run_id=run_id, description=f'Running the operator {self.name}')
@@ -101,8 +102,14 @@ def cache_operator_output(func):
         else:
             federated = None
 
-        if federated is not None and 'skip_operators' in federated and self.operator_out_dir in federated['skip_operators'] and last_round is False:
+        if federated is not None and 'skip_operators' in federated and self.operator_out_dir in federated['skip_operators']:
+            if last_round is False:
                 print('Skipping')
+                skip_downstream_tasks = [task for task in downstream_tasks if task.task_id in federated['skip_operators']]
+                self.skip(dag_run, dag_run.execution_date, skip_downstream_tasks)
+                return
+            elif not set(downstream_tasks_ids).issubset(set(federated['skip_operators'])):
+                print('Soft skipping, since we are in the last round...')
                 return
 
         if federated is not None and 'from_previous_dag_run' in federated and federated['from_previous_dag_run'] is not None:
@@ -140,28 +147,24 @@ def cache_operator_output(func):
         else:
             print('Caching is not used!')
 
-        if federated is not None and 'skip_operators' in federated and last_round is False:
-            downstream_tasks_ids = [task.task_id for task in downstream_tasks]
-            if downstream_tasks and set(downstream_tasks_ids).issubset(set(federated['skip_operators'])):
-                print('We will skip the rest!', downstream_tasks)
-                if 'before_previous_dag_run' in federated and federated['before_previous_dag_run'] is not None:
-                    before_previous_dag_run_dir = os.path.join(WORKFLOW_DIR, conf['federated_form']['before_previous_dag_run'])
-                    print(f'Removing batch files from before_previous_dag_run_dir: {before_previous_dag_run_dir}')
-                    if os.path.isdir(before_previous_dag_run_dir):
-                        shutil.rmtree(before_previous_dag_run_dir)     
-                print('Update remote job')
-                if conf is not None and 'client_job_id' in conf:
-                    update_job(conf['client_job_id'], status='finished', run_id=dag_run.run_id, description=f'Finished successfully')
+        if federated is not None and 'skip_operators' in federated and last_round is False and set(downstream_tasks_ids).issubset(set(federated['skip_operators'])):
+            print('The rest is skipped cleaning up!', downstream_tasks)
+            if 'before_previous_dag_run' in federated and federated['before_previous_dag_run'] is not None:
+                before_previous_dag_run_dir = os.path.join(WORKFLOW_DIR, conf['federated_form']['before_previous_dag_run'])
+                print(f'Removing batch files from before_previous_dag_run_dir: {before_previous_dag_run_dir}')
+                if os.path.isdir(before_previous_dag_run_dir):
+                    shutil.rmtree(before_previous_dag_run_dir)     
+            print('Update remote job')
+            if conf is not None and 'client_job_id' in conf:
+                update_job(conf['client_job_id'], status='finished', run_id=dag_run.run_id, description=f'Finished successfully')
 
-                if 'before_previous_dag_run' in federated and federated['before_previous_dag_run'] is not None:
-                    before_previous_dag_run_dir = os.path.join(WORKFLOW_DIR, conf['federated_form']['before_previous_dag_run'])
-                    print(f'Removing batch files from before_previous_dag_run_dir: {before_previous_dag_run_dir}')
-                    if os.path.isdir(before_previous_dag_run_dir):
-                        shutil.rmtree(before_previous_dag_run_dir)
-
-                print('Skipping the following tasks', downstream_tasks)
-                self.skip(dag_run, dag_run.execution_date, downstream_tasks)
-
+            if 'before_previous_dag_run' in federated and federated['before_previous_dag_run'] is not None:
+                before_previous_dag_run_dir = os.path.join(WORKFLOW_DIR, conf['federated_form']['before_previous_dag_run'])
+                print(f'Removing batch files from before_previous_dag_run_dir: {before_previous_dag_run_dir}')
+                if os.path.isdir(before_previous_dag_run_dir):
+                    shutil.rmtree(before_previous_dag_run_dir)
+            print('Skipping the following tasks', downstream_tasks)
+            self.skip(dag_run, dag_run.execution_date, downstream_tasks)
         return x
 
     return wrapper
